Reduced amino acid alphabets

The AA_REDUCED list contains reductions of the standard amino acid alphabet (AA_STANDARD).

Usage

AA_REDUCED

Format

An object of class list of length 160.

Details

Reduced amino alphabets can sometimes improve sensitivity and specificity of finding homologous matches between amino acid sequences. The first 12 AA_REDUCED alphabets were optimized for finding synteny between genomic sequences. The next 113 alphabets are from a review of published amino acid alphabets (Solis, 2015). The following 17 alphabets were optimized for amino acid classification. The final 18 alphabets are progressive mergers based on average similarities in PFASUM.

References

Solis, A. (2015). Amino acid alphabet reduction preserves fold information contained in contact interactions in proteins. Proteins: Structure, Function, and Genetics, 83(12), 2198-2216.

See also

Examples

str(AA_REDUCED)
#> List of 160
#>  $ : chr [1:16] "A" "C" "D" "E" ...
#>  $ : chr [1:14] "AG" "C" "DE" "FY" ...
#>  $ : chr [1:14] "AG" "C" "DN" "E" ...
#>  $ : chr [1:17] "A" "C" "D" "E" ...
#>  $ : chr [1:15] "AG" "C" "D" "E" ...
#>  $ : chr [1:14] "AG" "C" "D" "E" ...
#>  $ : chr [1:16] "A" "C" "D" "E" ...
#>  $ : chr [1:17] "A" "C" "D" "E" ...
#>  $ : chr [1:14] "AG" "C" "D" "EQ" ...
#>  $ : chr [1:13] "AG" "C" "DN" "EQ" ...
#>  $ : chr [1:16] "AS" "C" "D" "E" ...
#>  $ : chr [1:16] "A" "C" "D" "E" ...
#>  $ : chr [1:8] "AGILV" "CM" "DE" "FWY" ...
#>  $ : chr [1:4] "AFILMPVW" "CGNQSTY" "DE" "HKR"
#>  $ : chr [1:3] "ACGPSTWY" "DEHKNQR" "FILMV"
#>  $ : chr [1:10] "ACGS" "DE" "FWY" "HKR" ...
#>  $ : chr [1:2] "ACFILMVWY" "DEGHKNPQRST"
#>  $ : chr [1:3] "AFILMVWY" "C" "DEGHKNPQRST"
#>  $ : chr [1:4] "AFILMVWY" "C" "DEKR" "GHNPQST"
#>  $ : chr [1:5] "AFILMVWY" "C" "DE" "GHNPQST" ...
#>  $ : chr [1:6] "A" "C" "DE" "FILMVWY" ...
#>  $ : chr [1:7] "A" "C" "DE" "FILMVWY" ...
#>  $ : chr [1:8] "A" "C" "DE" "FILMV" ...
#>  $ : chr [1:9] "A" "C" "DE" "FILMV" ...
#>  $ : chr [1:10] "A" "C" "DE" "FILMV" ...
#>  $ : chr [1:10] "AST" "C" "DEN" "FY" ...
#>  $ : chr [1:2] "ACFGILMPSTVWY" "DEHKNQR"
#>  $ : chr [1:4] "AGPST" "CILMV" "DEHKNQR" "FWY"
#>  $ : chr [1:8] "AG" "CILMV" "DENQ" "FWY" ...
#>  $ : chr [1:10] "A" "C" "DENQ" "FWY" ...
#>  $ : chr [1:2] "ADEGHKNPQRST" "CFILMVWY"
#>  $ : chr [1:3] "ADEGHKNPQRST" "CFILMVY" "W"
#>  $ : chr [1:4] "ADEGHKNQRST" "CFILMVY" "P" "W"
#>  $ : chr [1:5] "ADEGKNQRST" "CFILMVY" "H" "P" ...
#>  $ : chr [1:6] "ADEGKNQRST" "C" "FILMVY" "H" ...
#>  $ : chr [1:7] "ADEKNQRST" "C" "FILMVY" "G" ...
#>  $ : chr [1:8] "AEKQRST" "C" "DN" "FILMVY" ...
#>  $ : chr [1:4] "ADEGKNPQRST" "C" "FILMVWY" "H"
#>  $ : chr [1:5] "ADEGKNPQRST" "C" "FILMVY" "H" ...
#>  $ : chr [1:6] "ADEGKNQRST" "C" "FILMVY" "H" ...
#>  $ : chr [1:7] "ADEKNQRST" "C" "FILMVY" "G" ...
#>  $ : chr [1:8] "ADEKNQRST" "C" "FILMV" "G" ...
#>  $ : chr [1:9] "ADNST" "C" "EKQR" "FILVY" ...
#>  $ : chr [1:2] "ACDEGHKNPQRST" "FILMVWY"
#>  $ : chr [1:3] "ADEGHKNPQRST" "C" "FILMVWY"
#>  $ : chr [1:4] "ADEGHKNPQRST" "C" "FILMVY" "W"
#>  $ : chr [1:5] "AGPST" "C" "DEHKNQR" "FILMVY" ...
#>  $ : chr [1:6] "AGPST" "C" "DEKNQR" "FILMVY" ...
#>  $ : chr [1:7] "ADGNST" "C" "EKQR" "FILMVY" ...
#>  $ : chr [1:8] "ADGNST" "C" "EKQR" "FY" ...
#>  $ : chr [1:9] "AGST" "C" "DN" "EKQR" ...
#>  $ : chr [1:10] "AGST" "C" "DN" "EQ" ...
#>  $ : chr [1:5] "ADEGHKNPQRST" "C" "FY" "ILMV" ...
#>  $ : chr [1:2] "ADEGHKNPQRST" "CFILMVWY"
#>  $ : chr [1:5] "APST" "CFWY" "DEHKNQR" "G" ...
#>  $ : chr [1:8] "AST" "CFWY" "DEQ" "G" ...
#>  $ : chr [1:10] "AST" "C" "DEQ" "FWY" ...
#>  $ : chr [1:10] "AC" "DE" "FWY" "G" ...
#>  $ : chr [1:6] "AST" "CP" "DEHKNQR" "FWY" ...
#>  $ : chr [1:8] "AST" "C" "DHN" "EKQR" ...
#>  $ : chr [1:10] "AST" "C" "DN" "EQ" ...
#>  $ : chr [1:10] "AST" "C" "DEN" "FY" ...
#>  $ : chr [1:2] "ADEGHKNPQRST" "CFILMVWY"
#>  $ : chr [1:3] "ACGPST" "DEHKNQR" "FILMVWY"
#>  $ : chr [1:4] "ACGST" "DEHKNQR" "FPWY" "ILMV"
#>  $ : chr [1:5] "ACST" "DEKNQR" "FHWY" "GP" ...
#>  $ : chr [1:6] "ACST" "DEKNQR" "FHWY" "G" ...
#>  $ : chr [1:7] "ACST" "DEN" "FWY" "G" ...
#>  $ : chr [1:8] "AST" "C" "DEN" "FWY" ...
#>  $ : chr [1:9] "AST" "C" "DEN" "FWY" ...
#>  $ : chr [1:10] "AST" "C" "DEN" "FY" ...
#>  $ : chr [1:5] "AG" "C" "DEKNPQRST" "FILMVWY" ...
#>  $ : chr [1:2] "ACEFHIKLMQRVWY" "DGNPST"
#>  $ : chr [1:3] "AEHKQR" "CFILMVWY" "DGNPST"
#>  $ : chr [1:4] "AEHKQR" "CFILMVWY" "DNST" "GP"
#>  $ : chr [1:5] "AEHKQR" "CST" "DN" "FILMVWY" ...
#>  $ : chr [1:6] "AEKQR" "CHST" "DN" "FIV" ...
#>  $ : chr [1:7] "ACH" "DN" "EKQR" "FIV" ...
#>  $ : chr [1:8] "AM" "CF" "DNS" "EKQR" ...
#>  $ : chr [1:9] "AMW" "CY" "DNS" "EKQR" ...
#>  $ : chr [1:10] "AM" "C" "DNS" "EKQR" ...
#>  $ : chr [1:2] "ACDEFHIKLMNQRSTVWY" "GP"
#>  $ : chr [1:3] "ACDEFHIKLMNQRSTVWY" "G" "P"
#>  $ : chr [1:4] "ADEKNQRST" "CFHILMVWY" "G" "P"
#>  $ : chr [1:5] "AEHKQRST" "CFILMVWY" "DN" "G" ...
#>  $ : chr [1:6] "AEFHIKLMQRVWY" "CT" "DN" "G" ...
#>  $ : chr [1:7] "AEFIKLMQRVWY" "CH" "DN" "G" ...
#>  $ : chr [1:8] "AEFIKLMQRVWY" "CH" "D" "G" ...
#>  $ : chr [1:9] "AEFIKLMQRVWY" "C" "D" "G" ...
#>  $ : chr [1:10] "AEFIKLMQRVW" "C" "D" "G" ...
#>  $ : chr [1:2] "AFILMVWY" "CDEGHKNPQRST"
#>  $ : chr [1:3] "AFILMVWY" "CDEHKNQRST" "GP"
#>  $ : chr [1:4] "ALM" "CDEHKNQRST" "FIVWY" "GP"
#>  $ : chr [1:5] "ALM" "CDHNST" "EKQR" "FIVWY" ...
#>  $ : chr [1:6] "ALM" "CHT" "DNS" "EKQR" ...
#>  $ : chr [1:7] "A" "CHT" "DNS" "EKQR" ...
#>  $ : chr [1:8] "A" "CHT" "DNS" "EKQR" ...
#>  $ : chr [1:9] "A" "CHT" "DNS" "EKQR" ...
#>  $ : chr [1:10] "A" "CH" "DNS" "EKQR" ...
#>   [list output truncated]
AA_REDUCED[[1]]
#>  [1] "A"    "C"    "D"    "E"    "FY"   "G"    "H"    "ILMV" "K"    "N"   
#> [11] "P"    "Q"    "R"    "S"    "T"    "W"