Pairwise Aligns Syntenic Blocks

Performs pairwise alignment of all blocks of synteny between sets of sequences.

Usage

AlignSynteny(
  synteny,
  dbFile,
  tblName = "Seqs",
  identifier = "",
  processors = 1,
  verbose = TRUE,
  ...
)

Arguments

synteny

An object of class ``Synteny''.

dbFile

A SQLite connection object or a character string specifying the path to the database file.

tblName

Character string specifying the table where the sequences are located that were used to create the object synteny.

identifier

Optional character string used to narrow the search results to those matching a specific identifier, or an integer sequence corresponding to indices of rownames(synteny). If "" (the default), then all identifiers are selected from synteny.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to display progress.

...

Further arguments to be passed directly to AlignProfiles, including perfectMatch, misMatch, gapPower, terminalGap, restrict, normPower, and substitutionMatrix.

Value

A list with elements for each pair of identifiers in synteny. Each list element contains a DNAStringSetList one pairwise alignment per syntenic block.

Details

AlignSynteny will extract all sequence regions belonging to syntenic blocks in synteny, and perform pairwise alignment with AlignProfiles. Hits are used to anchor the alignment such that only the regions between anchors are aligned.

See also

Author

Erik Wright eswright@pitt.edu

Examples

db <- system.file("extdata", "Influenza.sqlite", package="DECIPHER")
synteny <- FindSynteny(db, minScore=50)
#> ================================================================================
#> 
#> Time difference of 0.43 secs
#> 
DNA <- AlignSynteny(synteny, db)
#> ================================================================================
#> 
#> Time difference of 0.5 secs
#> 
names(DNA)
#>  [1] "H9N2/H5N1" "H9N2/H2N2" "H9N2/H7N9" "H9N2/H1N1" "H5N1/H2N2" "H5N1/H7N9"
#>  [7] "H5N1/H1N1" "H2N2/H7N9" "H2N2/H1N1" "H7N9/H1N1"
DNA[[1]] # the first set of pairwise alignments
#> DNAStringSetList of length 8
#> [["1"]] H9N2=GCAAAAGCAGGCAAACCATTTGAATGGATGTCAATCCGACTTTACTTTTCTTAAAAGTGCCAGC...
#> [["2"]] H9N2=CAAAAGCAGGTCAATTATATTCAATATGGAAAGAATAAAAGAACTAAGAAATTTGATGTCGCAA...
#> [["3"]] H9N2=AAAGCAGGTACTGATCCAAAATGGAAGACTTTGTGCGACAGTGCTTCAATCCAATGATTGTCGA...
#> [["4"]] H9N2=AGCAGGGTTAATAATCACTCACTGAGTGACATCAACATCATGGCGTCGCAAGGCACCAAACGAT...
#> [["5"]] H9N2=CAAAAGCAGGTAGATATTGAAAGATGAGTCTTCTAACCGAGGTCGAAACGTACGTTCTCTCTAT...
#> [["6"]] H9N2=GTGACAAAGACATAATGGATTCCAACACTGTGTCAAGCTTTCAGGTAGACTGCTTTCTTTGGCA...
#> [["7"]] H9N2=GAATGGTCCTACATCGTCGAAAGATCATCAGCTGTAAATGGAACGTGTTACCCTGGGAATGTAG...
#> [["8"]] H9N2=ATGAATCCAAATCAAAAGATAATAGCACTTGGCTCTGTTTCTATAACTATTGCGACAATATGTT...
DNA[[1]][[1]] # the first block of synteny between H9N2 & H5N1
#> DNAStringSet object of length 2:
#>     width seq                                               names               
#> [1]  2328 GCAAAAGCAGGCAAACCATTTGA...GCTTGTCCTTCATGAAAAAATGC H9N2
#> [2]  2328 GCAAAAGCAGGCAAACCATTTGA...GCTTGTCCTTCATGAAAAAATGC H5N1
unlist(DNA[[2]]) # a DNAStringSet of synteny between H9N2 & H2N2
#> DNAStringSet object of length 16:
#>      width seq                                              names               
#>  [1]  2328 GCAAAAGCAGGCAAACCATTTGA...CTTGTCCTTCATGAAAAAATGC 1.H9N2
#>  [2]  2328 GCAAAAGCAGGCAAACCATTTGA...CTTGTCCTTCATGAAAAAATGC 1.H2N2
#>  [3]  2339 CAAAAGCAGGTCAATTATATTCA...TAAAAACGACCTTGTTTCTACT 2.H9N2
#>  [4]  2339 CAAAAGCAGGTCAATTATATTCA...TAAAAACGACCTTGTTTCTACT 2.H2N2
#>  [5]  2225 AAAGCAGGTACTGATCCAAAATG...GTCCAAAAAAGTACCTTGTTTC 3.H9N2
#>  ...   ... ...
#> [12]  1007 CAAAAGCAGGTAGATATTGAAAG...TCAGCATAGAGTTGGAGTAAAA 6.H2N2
#> [13]   890 AGCAAAAGCAGGGTGACAAAGAC...AAAAAACACCCTTGTTTCTACT 7.H9N2
#> [14]   890 AGCAAAAGCAGGGTGACAAAGAC...AAAAAACACCCTTGTTTCTACT 7.H2N2
#> [15]  1129 CTTTTCGTGTGGGGCATACATCA...TGCAGATGCAACATTTGTATAT 8.H9N2
#> [16]  1129 CTGTACATTTGGGGGGTTCACCA...ATTAGGTGCAACATTTGCATTT 8.H2N2