Predict Thermodynamic Parameters of Probe/Target Sequence Pairs

Calculates the Gibbs free energy, formamide melt point, and hybridization efficiency of probe/target (DNA/RNA) pairs.

Usage

CalculateEfficiencyFISH(probe, target, temp, P, ions, FA, batchSize = 1000)

Arguments

probe

A DNAStringSet object or character vector with unaligned probe sequences in 5' to 3' orientation.

target

A DNAStringSet object, RNAStringSet, or character vector with unaligned target or non-target sequences in 5' to 3' orientation. The DNA base Thymine will be treated the same as Uracil.

temp

Numeric specifying the hybridization temperature, typically 46 degrees Celsius.

P

Numeric giving the molar concentration of probes during hybridization.

ions

Numeric giving the molar sodium equivalent ionic concentration. Values may range between 0.01M and 1M. Note that salt correction is not available for thermodynamic rules of RNA/RNA interactions, which were determined at 1 molar concentration.

FA

Numeric concentration (as percent v/v) of the denaturant formamide in the hybridization buffer.

batchSize

Integer specifying the number of probes to simulate hybridization per batch. See the Description section below.

Value

A matrix of predicted hybridization efficiency (HybEff), formamide melt point (FAm), and free energy (ddG1 and dG1) for each probe/target pair of sequences.

Details

Hybridization of pairwise probe/target (DNA/RNA) pairs is simulated in silico. Gibbs free energies are obtained from system calls to OligoArrayAux, which must be properly installed (see the Notes section below). Probe/target pairs are sent to OligoArrayAux in batches of batchSize, which prevents systems calls from being too many characters. Note that OligoArrayAux does not support degeneracy codes (non-base letters), although they are accepted without error. Any sequences with ambiguity should be expanded into multiple permutations with Disambiguate before input.

Note

The program OligoArrayAux (http://www.unafold.org/Dinamelt/software/oligoarrayaux.php) must be installed in a location accessible by the system. For example, the following code should print the installed OligoArrayAux version when executed from the R console:

system("hybrid-min -V")

References

ES Wright et al. (2014) "Automated Design of Probes for rRNA-Targeted Fluorescence In Situ Hybridization Reveals the Advantages of Using Dual Probes for Accurate Identification." Applied and Environmental Microbiology, doi:10.1128/AEM.01685-14.

See also

Author

Erik Wright eswright@pitt.edu

Examples

probe <- c("GGGCTTTCACATCAGACTTAAGAAACC", "CCCCACGCTTTCGCGCC")
target <- reverseComplement(DNAStringSet(probe))
# not run (must have OligoArrayAux installed first):
if (FALSE) CalculateEfficiencyFISH(probe, target, temp=46, P=250e-9, ions=1, FA=35)