Cluster Sequences By Distance

Groups the sequences into approximate clusters of similarity.

Usage

Clusterize(
  myXStringSet,
  cutoff = 0,
  method = "overlap",
  includeTerminalGaps = FALSE,
  penalizeGapLetterMatches = NA,
  minCoverage = 0.5,
  maxPhase1 = 400,
  maxPhase2 = 400,
  maxPhase3 = 400,
  maxAlignments = 100,
  rareKmers = 50,
  probability = 0.999,
  invertCenters = FALSE,
  singleLinkage = FALSE,
  alphabet = AA_REDUCED[[152]],
  processors = 1,
  verbose = TRUE
)

Arguments

myXStringSet

The (unaligned) DNAStringSet, RNAStringSet, or AAStringSet to cluster.

cutoff

A vector of maximum distances (approximately) separating sequences in the same cluster (i.e., 1 - similarities). Multiple cutoffs may be provided in ascending or descending order. (See details section below.)

method

Character string determining the region in which distance is calculated. This should be (an unambiguous abbreviation of) one of "overlap", "shortest", or "longest". The default method ("overlap") calculates distance from the overlapping region between terminal gaps when includeTerminalGaps is FALSE and the entire alignment otherwise. Setting method to "shortest" or "longest" will use the region between the start and end of the shortest or longest sequence, respectively, for each pairwise distance. The method is only applicable when includeTerminalGaps is TRUE.

includeTerminalGaps

Logical specifying whether or not to include terminal gaps ("-" characters) in the pairwise alignments into the calculation of distance.

penalizeGapLetterMatches

Logical specifying whether or not to consider gap-to-letter matches as mismatches. If FALSE, then gap-to-letter matches are not included in the total length used to calculate distance, and if TRUE all gaps-to-letter pairs are considered mismatches. The default (NA) is to penalize gap-to-letter mismatches once per insertion or deletion, which treats runs of gaps (i.e., indels) as equivalent to a single mismatch.

minCoverage

Numeric between zero and one giving the minimum fraction of sequence positions (not gap or mask) in the shortest sequence that must be overlapping with the longer sequence for the sequences to be clustered.

maxPhase1

An integer specifying the maximum number of passes through the sequences to perform in the initial partitioning of the sequences.

maxPhase2

An integer giving the maximum number of replicates to perform when ordering sequences based on their k-mer similarity.

maxPhase3

An integer determining the number of comparisons per sequence to perform when attempting to find cluster centers.

maxAlignments

An integer designating the maximum number of alignments to perform when attempting to assign a sequence to an existing cluster.

rareKmers

An integer setting the number of rare k-mers to record per sequence. Larger values require more memory but may improve accuracy with diminishing returns.

probability

Numeric between 0 and 1 (exclusive) defining the approximate probability of clustering sequences that are exactly cutoff distant. Typically near, but always less than, 1. Lower values result in faster clustering at the expense of effectiveness.

invertCenters

Logical controlling whether the cluster center is inverted (i.e., multiplied by -1), which allows the centers to be determined from the results. The default (FALSE) only returns positive cluster numbers. If TRUE, the center sequence(s) of each cluster are negative.

singleLinkage

Logical specifying whether to perform single-linkage clustering. The default (FALSE) only establishes linkage to the cluster center. Single-linkage clustering creates broader clusters that may better correspond to natural groups depending on the application.

alphabet

Character vector of amino acid groupings used to reduce the 20 standard amino acids into smaller groups. Alphabet reduction helps to find more distant homologies between sequences in phase 1.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to display progress.

Value

A data.frame is returned with dimensions \(N*M\), where each one of \(N\) sequences is assigned to a cluster at the \(M\)-level of cutoff. The row.names of the data.frame correspond to the names of myXStingSet.

Details

Clusterize groups the input sequences into approximate clusters using a heuristic algorithm with linear time and memory complexity. In phase 1, the sequences are partitioned into groups of similarity. In phase 2, the sequences are ordered by k-mer similarity by relatedness sorting. In phase 3, the sequences are iteratively clustered in this order by their similarity to surrounding sequences. That is, the first sequence becomes the representative of cluster #1. If the second sequence is within cutoff distance then it is added to the cluster, otherwise it becomes a new cluster representative. The remaining sequences are matched to cluster representatives in a similar fashion until all sequences belong to a cluster. In the majority of cases, this process results in clusters with members separated by less than cutoff distance, and all cluster members must be within cutoff distance of their cluster representative.

The calculation of distance can be controlled in multiple ways, with each parameterization of distance having advantages and disadvantages. By default, distance is the fraction of positions that are different, including gaps, within the overlapping region in a pairwise alignment. The defaults will handle partial-length sequences well, but also cluster sequences with high similarity between their opposite ends. For this reason, it is important to set minimumCoverage (by default 0.5) such that distances are based off of considerable overlap between sequences in the pairwise alignment. This distance parameterization works well, but there are reasonable alternatives.

If penalizeGapLetterMatches is FALSE, the distance will exclude gap regions. If includeTerminalGaps is TRUE, the calculation of distance will use the entire (global) alignment. If method is "shortest" and includeTerminalGaps is set to TRUE, then the distance is calculated for the region encompassed by the shorter sequence in each pair, which is the common definition of distance used by other clustering programs. This common definition of distance will more sometimes separate partly overlapping sequences, which is why it is not the default.

The algorithm requires time proportional to the number of input sequences in myXStringSet. The phase 1, up to maxPhase1 passes through the sequences are needed. In phase 2, the sequences are compared with up to maxPhase2 passes that each take linear time. Ordering of the sequences is performed in linear time using radix sorting. In phase 3, each sequence is compared with up to maxPhase3 previous cluster representatives of sequences sharing rareKmers or nearby sequences in the relatedness ordering. This is possible because the sequences are sorted by relatedness, such that more recent cluster representatives are more similar. Hence, the complete algorithm scales in linear time asymptotically and returns clusters of sequences within cutoff distance of their center sequence.

Multiple cutoffs can be provided in sorted order, which saves time because phases 1 and 2 only need to be performed once. If the cutoffs are provided in descending order then clustering at each new value of cutoff is continued within the prior cutoff's clusters. In this way clusters at lower values of cutoff are completely contained within their ``umbrella'' clusters at higher values of cutoff. This slightly accelerates the clustering process, because each subsequent group is only clustered within the previous group. If multiple cutoffs are provided in ascending order then clustering at each level of cutoff is independent of the prior level.

Note, all three phases of the algorithm are stochastic. Hence, clusters can vary from run-to-run unless the random number seed is set for repeatability (i.e., with set.seed). Also, invertCenters can be used to determine the center sequence of each cluster from the output. Since identical sequences will always be assigned the same cluster numbers, it is possible for more than one input sequence in myXStringSet to be assigned as the center of a cluster if they are identical.

See also

Author

Erik Wright eswright@pitt.edu

Examples

fas <- system.file("extdata", "50S_ribosomal_protein_L2.fas", package="DECIPHER")
dna <- readDNAStringSet(fas)
aa <- translate(dna)

# typical usage (e.g., clustering at >= 90 percent similarity)
clusters <- Clusterize(aa, cutoff=0.1) # set processors = NULL for max speed
#> Partitioning sequences by 6-mer similarity:
#> iteration 8 of up to 400 (100.0% coverage)
#> 
#> Time difference of 0.01 secs
#> 
#> Sorting by relatedness within 1 group:
#> iteration 51 of up to 51 (100.0% stability) 
#> 
#> Time difference of 0.48 secs
#> 
#> Clustering sequences by 4-mer similarity:
#> ================================================================================
#> 
#> Clusters via relatedness sorting: 100% (0% exclusively)
#> Clusters via rare 4-mers: 100% (0% exclusively)
#> Estimated clustering effectiveness: 100%
#> Time difference of 0.04 secs
#> 
head(clusters)
#>                                     cluster
#> Rickettsia prowazekii str. Dachau        77
#> Porphyromonas gingivalis W83             69
#> Porphyromonas gingivalis TDC60           69
#> Porphyromonas gingivalis ATCC 33277      69
#> Pasteurella multocida 671/90             76
#> Pasteurella multocida 36950              76

# typical usage (e.g., obtaining cluster representatives)
clusters <- Clusterize(aa, cutoff=0.1, invertCenters=TRUE)
#> Partitioning sequences by 6-mer similarity:
#> iteration 8 of up to 400 (100.0% coverage)
#> 
#> Time difference of 0.01 secs
#> 
#> Sorting by relatedness within 1 group:
#> iteration 51 of up to 51 (100.0% stability) 
#> 
#> Time difference of 0.47 secs
#> 
#> Clustering sequences by 4-mer similarity:
#> ================================================================================
#> 
#> Clusters via relatedness sorting: 100% (0% exclusively)
#> Clusters via rare 4-mers: 100% (0% exclusively)
#> Estimated clustering effectiveness: 100%
#> Time difference of 0.04 secs
#> 
aa[clusters[[1]] < 0]
#> AAStringSet object of length 268:
#>       width seq                                             names               
#>   [1]   274 MGIRKLKPTTPGQRHKVIGAFD...KGLKTRAPKKHSSKYIIERRKK Porphyromonas gin...
#>   [2]   274 MGIRKLKPTTPGQRHKVIGAFD...KGLKTRAPKKHSSKYIIERRKK Porphyromonas gin...
#>   [3]   274 MGIRKLKPTTPGQRHKVIGAFD...KGLKTRAPKKHSSKYIIERRKK Porphyromonas gin...
#>   [4]   279 VGIKKYKPTTNGRRNMTASDFS...KTRNKKARSNKLIVRGRRPGKH Lactobacillus pla...
#>   [5]   279 VGIKKYKPTTNGRRNMTASDFS...KTRNKKARSNKLIVRGRRPGKH Lactobacillus pla...
#>   ...   ... ...
#> [264]   281 MFKKYRPVTPGTRQLALPTNEQ...KGLVTRSNKKSNKLIVKRRRKK Candidatus Protoc...
#> [265]   274 MPIQKCKPTSPGRRFVEKVVHS...KGYKTRTNKRTTKMIIRDRRVK Acinetobacter sp....
#> [266]   288 MGIRVYRPYTPGVRQKTVSDFA...SDALIVRRRKKSSKRGRGGRQS Thermosynechococc...
#> [267]   277 MALKTFNPTTPGQRQLVMVDRS...KKTRSNKSTNKFILLSRHKRKK Bradyrhizobium ja...
#> [268]   280 MGIRKYRPMTPGTRQRSGADFA...RKRRKPSSKFIIRRRKTASGRG Gloeobacter viola...

# cluster each cutoff within the previous cluster (slightly faster)
clusters <- Clusterize(aa, cutoff=seq(0.7, 0, -0.1))
#> Partitioning sequences by 6-mer similarity:
#> iteration 8 of up to 400 (100.0% coverage)
#> 
#> Time difference of 0.01 secs
#> 
#> Sorting by relatedness within 1 group:
#> iteration 51 of up to 51 (100.0% stability) 
#> 
#> Time difference of 0.63 secs
#> 
#> Clustering sequences by 4-mer similarity:
#> ================================================================================
#> 
#> Clusters via relatedness sorting: 100% (0% exclusively)
#> Clusters via rare 4-mers: 100% (0% exclusively)
#> Estimated clustering effectiveness: 100%
#> Time difference of 0.08 secs
#> 
head(clusters)
#>                                     cluster_0_7 cluster_0_6 cluster_0_5
#> Rickettsia prowazekii str. Dachau             1           1           1
#> Porphyromonas gingivalis W83                  1           1           1
#> Porphyromonas gingivalis TDC60                1           1           1
#> Porphyromonas gingivalis ATCC 33277           1           1           1
#> Pasteurella multocida 671/90                  1           1           1
#> Pasteurella multocida 36950                   1           1           1
#>                                     cluster_0_4 cluster_0_3 cluster_0_2
#> Rickettsia prowazekii str. Dachau            19           3          64
#> Porphyromonas gingivalis W83                 12          20          35
#> Porphyromonas gingivalis TDC60               12          20          35
#> Porphyromonas gingivalis ATCC 33277          12          20          35
#> Pasteurella multocida 671/90                 15           8          56
#> Pasteurella multocida 36950                  15           8          56
#>                                     cluster_0_1 cluster_0
#> Rickettsia prowazekii str. Dachau             3       100
#> Porphyromonas gingivalis W83                 48        44
#> Porphyromonas gingivalis TDC60               48        44
#> Porphyromonas gingivalis ATCC 33277          48        44
#> Pasteurella multocida 671/90                 15        84
#> Pasteurella multocida 36950                  15        84
apply(clusters, 2, max) # number of clusters per cutoff
#> cluster_0_7 cluster_0_6 cluster_0_5 cluster_0_4 cluster_0_3 cluster_0_2 
#>           1           1           4          20          44          66 
#> cluster_0_1   cluster_0 
#>          88         102 

# cluster each cutoff independently (possibly fewer clusters per cutoff)
clusters <- Clusterize(aa, cutoff=seq(0, 0.7, 0.1))
#> Partitioning sequences by 6-mer similarity:
#> iteration 8 of up to 400 (100.0% coverage)
#> 
#> Time difference of 0.01 secs
#> 
#> Sorting by relatedness within 1 group:
#> iteration 51 of up to 51 (100.0% stability) 
#> 
#> Time difference of 0.64 secs
#> 
#> Clustering sequences by 4-mer similarity:
#> ================================================================================
#> 
#> Clusters via relatedness sorting: 100% (0% exclusively)
#> Clusters via rare 4-mers: 100% (0% exclusively)
#> Estimated clustering effectiveness: 100%
#> Time difference of 0.23 secs
#> 
head(clusters)
#>                                     cluster_0 cluster_0_1 cluster_0_2
#> Rickettsia prowazekii str. Dachau          99          78          58
#> Porphyromonas gingivalis W83               82          70          53
#> Porphyromonas gingivalis TDC60             82          70          53
#> Porphyromonas gingivalis ATCC 33277        82          70          53
#> Pasteurella multocida 671/90               94          77          52
#> Pasteurella multocida 36950                94          77          52
#>                                     cluster_0_3 cluster_0_4 cluster_0_5
#> Rickettsia prowazekii str. Dachau            34          15           4
#> Porphyromonas gingivalis W83                 31           7           4
#> Porphyromonas gingivalis TDC60               31           7           4
#> Porphyromonas gingivalis ATCC 33277          31           7           4
#> Pasteurella multocida 671/90                 25          10           4
#> Pasteurella multocida 36950                  25          10           4
#>                                     cluster_0_6 cluster_0_7
#> Rickettsia prowazekii str. Dachau             1           1
#> Porphyromonas gingivalis W83                  1           1
#> Porphyromonas gingivalis TDC60                1           1
#> Porphyromonas gingivalis ATCC 33277           1           1
#> Pasteurella multocida 671/90                  1           1
#> Pasteurella multocida 36950                   1           1
apply(clusters, 2, max) # number of clusters per cutoff
#>   cluster_0 cluster_0_1 cluster_0_2 cluster_0_3 cluster_0_4 cluster_0_5 
#>         102          84          61          35          16           4 
#> cluster_0_6 cluster_0_7 
#>           1           1 

# make cluster center(s) negative for tracking
clusters <- Clusterize(aa, cutoff=0.5, invertCenters=TRUE)
#> Partitioning sequences by 6-mer similarity:
#> iteration 8 of up to 400 (100.0% coverage)
#> 
#> Time difference of 0.01 secs
#> 
#> Sorting by relatedness within 1 group:
#> iteration 20 of up to 51 (100.0% stability) 
#> 
#> Time difference of 0.25 secs
#> 
#> Clustering sequences by 4-mer similarity:
#> ================================================================================
#> 
#> Clusters via relatedness sorting: 100% (0% exclusively)
#> Clusters via rare 4-mers: 100% (0% exclusively)
#> Estimated clustering effectiveness: 100%
#> Time difference of 0.03 secs
#> 
head(clusters)
#>                                     cluster
#> Rickettsia prowazekii str. Dachau         4
#> Porphyromonas gingivalis W83              4
#> Porphyromonas gingivalis TDC60            4
#> Porphyromonas gingivalis ATCC 33277       4
#> Pasteurella multocida 671/90              4
#> Pasteurella multocida 36950               4
clusters[clusters$cluster < 0,, drop=FALSE]
#>                                                      cluster
#> Mycoplasma pneumoniae M129-B7                             -3
#> Aquifex aeolicus VF5                                      -1
#> Sinorhizobium meliloti AK83                               -4
#> Sinorhizobium meliloti BL225C                             -4
#> Sinorhizobium meliloti GR4                                -4
#> Sinorhizobium meliloti 2011                               -4
#> Sinorhizobium medicae WSM419                              -4
#> Sinorhizobium meliloti SM11                               -4
#> Prochlorococcus marinus subsp. marinus str. CCMP1375      -2
unique(aa[clusters$cluster < 0]) # unique cluster centers
#> AAStringSet object of length 4:
#>     width seq                                               names               
#> [1]   287 MPIKKIISRSNSGIHHSTVIDYK...VKTRNMKKASTNLIIRNRKGEQY Mycoplasma pneumo...
#> [2]   304 MGVRKLKPVTNGTRHAVLYDFEE...YKTRRGKKYSDQFIVTRRDGRPL Aquifex aeolicus VF5
#> [3]   278 MALKSFNPTTPSQRQLVIVSRAG...GKRTRSNKSTDKFIMRSRHQRKK Sinorhizobium mel...
#> [4]   287 MAIRTFRPYTPGTRTRVVTDFNE...PSNRFVLRKRRRVSKRSRGGRDS Prochlorococcus m...
apply(clusters, 2, function(x) max(abs(x))) # number of clusters
#> cluster 
#>       4 

# cluster nucleotide sequences
clusters <- Clusterize(dna, cutoff=0.5, invertCenters=TRUE)
#> Partitioning sequences by 9-mer similarity:
#> iteration 96 of up to 400 (100.0% coverage)
#> 
#> Time difference of 0.1 secs
#> 
#> Sorting by relatedness within 89 groups:
#> iteration 37 of up to 37 (100.0% stability) 
#> 
#> Time difference of 0.23 secs
#> 
#> Clustering sequences by 9-mer similarity:
#> ================================================================================
#> 
#> Clusters via relatedness sorting: 100% (0% exclusively)
#> Clusters via rare 6-mers: 100% (0% exclusively)
#> Estimated clustering effectiveness: 100%
#> Time difference of 0.42 secs
#> 
head(clusters)
#>                                     cluster
#> Rickettsia prowazekii str. Dachau        -1
#> Porphyromonas gingivalis W83             -2
#> Porphyromonas gingivalis TDC60            2
#> Porphyromonas gingivalis ATCC 33277       2
#> Pasteurella multocida 671/90              2
#> Pasteurella multocida 36950               2
apply(clusters, 2, function(x) max(abs(x))) # number of clusters
#> cluster 
#>       4