Expand Ambiguities into All Permutations of a DNAStringSet

Performs the inverse function of ConsensusSequence by expanding any ambiguities present in sequences.

Usage

Disambiguate(myXStringSet)

Arguments

myXStringSet

A DNAStringSet or RNAStringSet object of sequences.

Value

A DNAStringSetList or RNAStringSetList with one element for each sequence in myXStringSet.

Details

Ambiguity codes in the IUPAC_CODE_MAP can be used to represent multiple nucleotides at a single position. Using these letters, multiple oligonucleotide permutations can be represented with a single ambiguous sequence. This function expands each sequence in the DNAStringSet input into all of its permutations. Note that sequences with many ambiguities can result in a very large number of potential permutations.

See also

Author

Erik Wright eswright@pitt.edu

Examples

dna <- DNAStringSet(c("ACST", "NNN"))
dna_list <- Disambiguate(dna)
dna_list[[1]]
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     4 ACCT
#> [2]     4 ACGT
dna_list[[2]]
#> DNAStringSet object of length 64:
#>      width seq
#>  [1]     3 AAA
#>  [2]     3 CCC
#>  [3]     3 GGG
#>  [4]     3 TTT
#>  [5]     3 ACA
#>  ...   ... ...
#> [60]     3 TCG
#> [61]     3 ATT
#> [62]     3 CAA
#> [63]     3 GCC
#> [64]     3 TGG
unlist(dna_list)
#> DNAStringSet object of length 66:
#>      width seq
#>  [1]     4 ACCT
#>  [2]     4 ACGT
#>  [3]     3 AAA
#>  [4]     3 CCC
#>  [5]     3 GGG
#>  ...   ... ...
#> [62]     3 TCG
#> [63]     3 ATT
#> [64]     3 CAA
#> [65]     3 GCC
#> [66]     3 TGG

rna <- RNAStringSet(c("ACGU", "AGAU")) # 2 permutations
rna <- ConsensusSequence(rna) # "ASRU"
Disambiguate(rna) # 4 permutations
#> RNAStringSetList of length 1
#> [[1]] ACAU AGGU ACGU AGAU