Extract Predicted Genes from a Genome

Extracts predicted genes from the genome used for prediction.

Usage

ExtractGenes(x, myDNAStringSet, type = "DNAStringSet", ...)

Arguments

x

An object of class Genes.

myDNAStringSet

The DNAStringSet object used in generating x.

type

The class of sequences to return. This should be (an unambiguous abbreviation of) one of "AAStringSet", "DNAStringSet" (the default), or "RNAStringSet".

...

Other parameters passed directly to translate.

Value

An "AAStringSet", "DNAStringSet", or "RNAStringSet" determined by type.

Details

Extracts a set of gene predictions as either DNA, mRNA, or proteins.

See also

Author

Erik Wright eswright@pitt.edu

Examples

# import a test genome
fas <- system.file("extdata",
  "Chlamydia_trachomatis_NC_000117.fas.gz",
  package="DECIPHER")
genome <- readDNAStringSet(fas)

x <- FindGenes(genome)
#> Iter  Models Start Motif  Init  Fold UpsNt  Term   RBS  Auto  Stop Genes
#> 1          1 18.40                 _                                 8861          1 18.40                 \                                 8861          1 18.40                 |                                 8861          1 18.40                 /                                 8861          1 18.40                 _                                 8861          1 18.40                 \                                 8861          1 18.40                 |                                 8861          1 18.40                 /                                 8861          1 18.40                 _                                 8862         12 18.40  0.48  0.90                                       8883         15 18.38  0.53  1.16  0.40  1.41                           8904         16 15.59  0.52  1.26  0.48  1.76  0.24                     8935         12 15.62  0.51  1.25  0.48  1.84  0.25  0.98  0.10  0.07   8976         13 15.61  0.51  1.25  0.49  1.91  0.25  1.05  0.10  0.07   8977         13 15.59  0.51  1.26  0.50  1.93  0.26  1.01  0.10  0.07   8988         10 15.59  0.51  1.26  0.52  1.94  0.26  1.04  0.10  0.07   8979         10 15.59  0.51  1.26  0.51  1.94  0.26  1.06  0.10  0.07   897
#> 
#> Time difference of 22.28 secs
#> 
genes <- ExtractGenes(x, genome)
proteins <- ExtractGenes(x, genome, type="AAStringSet")