Forms Groups By Rank

Agglomerates sequences into groups within a specified size range based on taxonomic rank.

Usage

FormGroups(
  dbFile,
  tblName = "Seqs",
  goalSize = 50,
  minGroupSize = 25,
  maxGroupSize = 5000,
  includeNames = FALSE,
  add2tbl = FALSE,
  verbose = TRUE
)

Arguments

dbFile

A SQLite connection object or a character string specifying the path to the database file.

tblName

Character string specifying the table where the rank information is located.

goalSize

Number of sequences required in each group to stop adding more sequences.

minGroupSize

Minimum number of sequences in each group required to stop trying to recombine with a larger group.

maxGroupSize

Maximum number of sequences in each group allowed to continue agglomeration.

includeNames

Logical indicating whether to include the formal scientific name in the group name.

add2tbl

Logical or a character string specifying the table name in which to add the result.

verbose

Logical indicating whether to display progress.

Value

A data.frame with the rank and corresponding group name as identifier. Note that quotes are stripped from group names to prevent problems that they may cause. The origin gives the rank preceding the identifier. The count denotes number of sequences corresponding to each rank. If add2tbl is not FALSE then the identifier'' and origin'' columns are updated in dbFile.

Details

FormGroups uses the ``rank'' field in the dbFile table to group sequences with similar taxonomic rank. Rank information must be present in the tblName, such as that created by default when importing sequences from a GenBank formatted file.

Rank information contains the formal scientific name on the first line, followed by the taxonomic lineage on subsequent lines. When includeNames is TRUE the formal scientific name is appended to the end of the group name, otherwise only the taxonomic lineage is used as the group name.

The algorithm ascends the taxonomic tree, agglomerating taxa into groups until the goalSize is reached. If the group size is below minGroupSize then further agglomeration is attempted with a larger group. If additional agglomeration results in a group larger than maxGroupSize then the agglomeration is undone so that the group is smaller. Setting minGroupSize to goalSize avoids the creation of polyphyletic groups. Note that this approach may often result in paraphyletic groups.

See also

Author

Erik Wright eswright@pitt.edu

Examples

db <- system.file("extdata", "Bacteria_175seqs.sqlite", package="DECIPHER")
g <- FormGroups(db, goalSize=10, minGroupSize=5, maxGroupSize=20)
#> ================================================================================
#> Formed 14 distinct groups.
#> Time difference of 0.01 secs
#> 
head(g)
#>                                                                                                                                                       rank
#> 15                                               uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Bacilli"; Bacillales; Bacillaceae; \nLysinibacillus.
#> 46                                       uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Bacilli"; Bacillales; \n"Staphylococcaceae"; Staphylococcus.
#> 14                                uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Bacilli"; "Lactobacillales"; \n"Carnobacteriaceae"; Dolosigranulum.
#> 16                 uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Clostridia"; Clostridiales; \nPeptococcaceae; Peptococcaceae 1; Desulfosporosinus.
#> 58 uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Erysipelotrichi"; \n"Erysipelotrichales"; Erysipelotrichaceae; \nunclassified_Erysipelotrichaceae.
#> 52                            uncultured bacterium\nRoot; Bacteria; "Bacteroidetes"; "Bacteroidia"; "Bacteroidales"; \n"Porphyromonadaceae"; Dysgonomonas.
#>    count        origin      identifier
#> 15     1 Root;Bacteria    _Firmicutes_
#> 46     2 Root;Bacteria    _Firmicutes_
#> 14     1 Root;Bacteria    _Firmicutes_
#> 16     1 Root;Bacteria    _Firmicutes_
#> 58     4 Root;Bacteria    _Firmicutes_
#> 52     3 Root;Bacteria _Bacteroidetes_
tapply(g$count, g$identifier, sum)
#>               Alphaproteobacteria                          Bacteria 
#>                                12                                13 
#>                Betaproteobacteria               Gammaproteobacteria 
#>                                20                                16 
#>                 Hyphomicrobiaceae                    Mycobacterium_ 
#>                                 7                                22 
#>                   Pseudomonadales                       Rhizobiales 
#>                                 9                                15 
#>                 Sphingomonadaceae                     Sphingomonas_ 
#>                                13                                 9 
#>                   _Bacteroidetes_                      _Firmicutes_ 
#>                                18                                 9 
#>                      _Nitrospira_ unclassified_Alphaproteobacteria_ 
#>                                 5                                 7