Determine the Number of Characters in Each Sequence

Counts the number of standard and non-standard characters in each sequence.

Usage

IdLengths(
  dbFile,
  tblName = "Seqs",
  type = "DNAStringSet",
  add2tbl = FALSE,
  batchSize = 10000,
  processors = 1,
  verbose = TRUE
)

Arguments

dbFile

A SQLite connection object or a character string specifying the path to the database file.

tblName

Character string specifying the table where the sequences are located.

type

The type of XStringSet being processed. This should be (an abbreviation of) one of "AAStringSet", "DNAStringSet", or "RNAStringSet".

add2tbl

Logical or a character string specifying the table name in which to add the result.

batchSize

Integer specifying the number of sequences to process at a time.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to display progress.

Value

A data.frame with the number of standard characters, nonstandard characters, and width of each sequence. The row.names of the data.frame correspond to the "row_names" in the tblName of the dbFile.

Details

IdLengths is designed to efficiently determine the number of standard and non-standard characters in every sequence within a database. Standard and non-standard characters are defined with respect to the type of the sequences. For DNA and RNA sequences there are four standard characters and 11 non-standard characters (i.e., ambiguity codes). For amino acid sequences there are 20 standard and seven non-standard characters (including stops). Gap (-''), missing (.''), and mask (``+'') characters count toward the width but not the number of standard or non-standard characters.

References

ES Wright (2016) "Using DECIPHER v2.0 to Analyze Big Biological Sequence Data in R". The R Journal, 8(1), 352-359.

See also

Author

Erik Wright eswright@pitt.edu

Examples

db <- system.file("extdata", "Bacteria_175seqs.sqlite", package="DECIPHER")
l <- IdLengths(db)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Lengths counted for 175 sequences.
#> 
#> Time difference of 0.02 secs
#> 
head(l)
#>   standard nonstandard width
#> 1     1222          13  1596
#> 2     1346           5  1596
#> 3     1323           3  1596
#> 4     1337           8  1596
#> 5     1318          25  1596
#> 6     1307           7  1596