Identify By Taxonomic Rank

Identifies sequences by a specific level of their taxonomic rank.

Usage

IdentifyByRank(
  dbFile,
  tblName = "Seqs",
  level = 0,
  add2tbl = FALSE,
  verbose = TRUE
)

Arguments

dbFile

A SQLite connection object or a character string specifying the path to the database file.

tblName

Character string specifying the table where the rank information is located.

level

Level of the taxonomic rank. (See details section below.)

add2tbl

Logical or a character string specifying the table name in which to add the result.

verbose

Logical indicating whether to print database queries and other information.

Value

A data.frame with the rank and corresponding identifier as identifier. Note that quotes are stripped from identifiers to prevent problems that they may cause. The origin

gives the rank preceding the identifier. If add2tbl is not FALSE then the ``identifier'' column is updated in dbFile.

Details

IdentifyByRank simply identifies a sequence by a specific level of its taxonomic rank. Requires that rank information be present in the tblName, such as that created by default when importing sequences from a GenBank formatted file.

The input parameter level should be an integer giving the ``level'' of the taxonomic rank to choose as the identifier. Negative levels are interpreted as being that many levels from the last level in each rank. The level zero selects the base level (see below).

If the specified level of rank does not exist then the closest rank is chosen. Therefore, setting level to Inf will always select the last taxonomic level (i.e., genus).

For example, a representative ``rank'' imported from a GenBank file is:
Saccharomyces cerevisiae
Eukaryota; Fungi; Ascomycota; Saccharomycotina; Saccharomycetes;
Saccharomycetales; Saccharomycetaceae; Saccharomyces.

Setting level to 0 would result in an identifier of Saccharomyces cerevisiae'', because it is on the first line. A \code{level} of \code{2} would return Fungi'', and -2 (second to last) would return Saccharomycetaceae''. A \code{level} of \code{Inf} would find the nearest level to the end, Saccharomyces''.

See also

Author

Erik Wright eswright@pitt.edu

Examples

db <- system.file("extdata", "Bacteria_175seqs.sqlite", package="DECIPHER")
ids <- IdentifyByRank(db, level=Inf)
#> 
#> Formed 72 distinct groups.
#> Time difference of 0.01 secs
#> 
head(ids)
#>                                                                                                                                                      rank
#> 1                                               uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Bacilli"; Bacillales; Bacillaceae; \nLysinibacillus.
#> 2                                       uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Bacilli"; Bacillales; \n"Staphylococcaceae"; Staphylococcus.
#> 3                                uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Bacilli"; "Lactobacillales"; \n"Carnobacteriaceae"; Dolosigranulum.
#> 4                 uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Clostridia"; Clostridiales; \nPeptococcaceae; Peptococcaceae 1; Desulfosporosinus.
#> 5 uncultured bacterium\nRoot; Bacteria; "Firmicutes"; "Erysipelotrichi"; \n"Erysipelotrichales"; Erysipelotrichaceae; \nunclassified_Erysipelotrichaceae.
#> 6                            uncultured bacterium\nRoot; Bacteria; "Bacteroidetes"; "Bacteroidia"; "Bacteroidales"; \n"Porphyromonadaceae"; Dysgonomonas.
#>                                                                                          origin
#> 1                             Root; Bacteria; "Firmicutes"; "Bacilli"; Bacillales; Bacillaceae;
#> 2                    Root; Bacteria; "Firmicutes"; "Bacilli"; Bacillales;  "Staphylococcaceae";
#> 3             Root; Bacteria; "Firmicutes"; "Bacilli"; "Lactobacillales";  "Carnobacteriaceae";
#> 4 Root; Bacteria; "Firmicutes"; "Clostridia"; Clostridiales;  Peptococcaceae; Peptococcaceae 1;
#> 5  Root; Bacteria; "Firmicutes"; "Erysipelotrichi";  "Erysipelotrichales"; Erysipelotrichaceae;
#> 6       Root; Bacteria; "Bacteroidetes"; "Bacteroidia"; "Bacteroidales";  "Porphyromonadaceae";
#>                         identifier
#> 1                   Lysinibacillus
#> 2                   Staphylococcus
#> 3                   Dolosigranulum
#> 4                Desulfosporosinus
#> 5 unclassified_Erysipelotrichaceae
#> 6                     Dysgonomonas