NonCoding Objects and Methods

Non-coding RNAs can be represented by their conserved sequence motifs, secondary structure, and k-mer frequencies. Class NonCoding provides objects and functions for representing non-coding RNAs.

Non-coding RNAs can be represented by their conserved sequence motifs, secondary structure, and k-mer frequencies. Class NonCoding provides objects and functions for representing non-coding RNAs.

Arguments

x

An object of class NonCoding.

...

Other optional parameters.

Details

Objects of class NonCoding are stored as lists containing a compact representation of a family of non-coding RNAs. The first list component is a matrix of sequence motifs that identify the non-coding RNAs, the second is a matrix of hairpin loops that are conserved across the family, the third is a list of k-mer frequencies derived from representative sequences, and the fourth is a vector of log-odds scores for sequence lengths. An optional fifth list component denotes the log-odds scores for dependencies among patterns. Patterns are defined by their distance to either end of the non-coding RNA, which helps to identify the boundaries of the non-coding RNA in a genome.

Objects of class NonCoding are stored as lists containing a compact representation of a family of non-coding RNAs. The first list component is a matrix of sequence motifs that identify the non-coding RNAs, the second is a matrix of hairpin loops that are conserved across the family, the third is a list of k-mer frequencies derived from representative sequences, and the fourth is a vector of log-odds scores for sequence lengths. An optional fifth list component denotes the log-odds scores for dependencies among patterns. Patterns are defined by their distance to either end of the non-coding RNA, which helps to identify the boundaries of the non-coding RNA in a genome.

References

Wright, E. S. (2021). FindNonCoding: rapid and simple detection of non-coding RNAs in genomes. Bioinformatics. https://doi.org/10.1093/bioinformatics/btab708

Wright, E. S. (2021). FindNonCoding: rapid and simple detection of non-coding RNAs in genomes. Bioinformatics. https://doi.org/10.1093/bioinformatics/btab708

See also

Author

Erik Wright eswright@pitt.edu

Examples

data(NonCodingRNA_Bacteria)
x <- NonCodingRNA_Bacteria
print(x)
#> $`tRNA-Ala`
#> NonCoding object with 10 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Arg`
#> NonCoding object with 7 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-Asn`
#> NonCoding object with 9 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Asp`
#> NonCoding object with 9 motifs, 2 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Cys`
#> NonCoding object with 8 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Gln`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Glu`
#> NonCoding object with 8 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Gly`
#> NonCoding object with 9 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-His`
#> NonCoding object with 10 motifs, 2 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Ile`
#> NonCoding object with 8 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Leu`
#> NonCoding object with 11 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-Lys`
#> NonCoding object with 9 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Met`
#> NonCoding object with 11 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Phe`
#> NonCoding object with 9 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Pro`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Ser`
#> NonCoding object with 10 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-Thr`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Trp`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Tyr`
#> NonCoding object with 14 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Val`
#> NonCoding object with 8 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Sec`
#> NonCoding object with 20 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`rRNA_5S-RF00001`
#> NonCoding object with 20 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`rRNA_16S-RF00177`
#> NonCoding object with 8 motifs, 20 hairpins, and 4-mer frequencies.
#> 
#> $`rRNA_23S-RF02541`
#> NonCoding object with 11 motifs, 20 hairpins, and 4-mer frequencies.
#> 
#> $`tmRNA-RF00023`
#> NonCoding object with 20 motifs, 6 hairpins, and 4-mer frequencies.
#> 
#> $`tmRNA_Alpha-RF01849`
#> NonCoding object with 17 motifs, 10 hairpins, and 4-mer frequencies.
#> 
#> $`RNase_P_class_A-RF00010`
#> NonCoding object with 20 motifs, 13 hairpins, and 4-mer frequencies.
#> 
#> $`RNase_P_class_B-RF00011`
#> NonCoding object with 15 motifs, 13 hairpins, and 4-mer frequencies.
#> 
#> $`SsrS-RF00013`
#> NonCoding object with 14 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`Intron_Gp_I-RF00028`
#> NonCoding object with 15 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`Intron_Gp_II-RF00029`
#> NonCoding object with 9 motifs, 2 hairpins, and 4-mer frequencies.
#> 
#> $`SmallSRP-RF00169`
#> NonCoding object with 16 motifs, 5 hairpins, and 4-mer frequencies.
#> 
#> $`Cyclic-di-GMP_Riboswitch-RF01051`
#> NonCoding object with 14 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`Cyclic-di-AMP_Riboswitch-RF00379`
#> NonCoding object with 19 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`T-box_Leader-RF00230`
#> NonCoding object with 20 motifs, 7 hairpins, and 4-mer frequencies.
#> 
#> $`Ribosomal_Protein_L10_Leader-RF00557`
#> NonCoding object with 11 motifs, 2 hairpins, and 4-mer frequencies.
#> 
#> $`Cobalamin_Riboswitch-RF00174`
#> NonCoding object with 11 motifs, 7 hairpins, and 4-mer frequencies.
#> 
#> $`TPP_Riboswitch-RF00059`
#> NonCoding object with 10 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`SAM_Riboswitch-RF00162`
#> NonCoding object with 14 motifs, 5 hairpins, and 4-mer frequencies.
#> 
#> $`Fluoride_Riboswitch-RF01734`
#> NonCoding object with 9 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`FMN_Riboswitch-RF00050`
#> NonCoding object with 12 motifs, 6 hairpins, and 4-mer frequencies.
#> 
#> $`Glycine_Riboswitch-RF00504`
#> NonCoding object with 10 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`HEARO-RF02033`
#> NonCoding object with 17 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`Flavo_1-RF01705`
#> NonCoding object with 8 motifs, 2 hairpins, and 4-mer frequencies.
#> 
#> $`Acido_Lenti_1-RF01687`
#> NonCoding object with 6 motifs, 2 hairpins, and 3-mer frequencies.
#> 
#> $`5'_ureB-RF02514`
#> NonCoding object with 10 motifs, 1 hairpin, and 4-mer frequencies.
#> 
class(x)
#> [1] "list"
attributes(x[[1]])
#> $names
#> [1] "motifs"       "hairpins"     "kmers"        "lengthScores" "dependence"  
#> 
#> $class
#> [1] "NonCoding"
#> 
#> $K
#> [1] 3
#> 
#> $minLength
#> [1] 47
#> 
#> $maxLength
#> [1] 154
#> 
#> $maxLoopLength
#> [1] 500
#> 
#> $background
#> meanlog   sdlog 
#>      NA      NA 
#> 
x[[1]] # the first non-coding RNA
#> NonCoding object with 10 motifs, 4 hairpins, and 3-mer frequencies.
x[[1]][["motifs"]] # sequence motifs
#>    begin_low begin_high end_low end_high          motif          pwm
#> 1          0          0      66       68       GGGGcydT 6.190340....
#> 2          8          8      58       60       AGCTCAGy 0.887309....
#> 3         16         16      55       57            TGG 0.005205....
#> 4         19         20      43       45   kAGAGCrCyTsc 0.024808....
#> 5         31         32      29       31 wTkGCAwksArGRG 0.261497....
#> 6         45         46      21       23       GTCrgsGG 0.016786....
#> 7         53         54      12       14      TTCGAwyCC 6.189574....
#> 8         62         63       1        3    scThrkCTCCA 6.190340....
#> 9         73         73       0        2              C 0.011661....
#> 10        74         74       0        0             CA 0.024090....
#>        minscore   prevalence   background
#> 1  0, 6.546.... 0.094171.... 0.990257....
#> 2  0, 7.992.... 0.122470.... 0.997144....
#> 3        0, Inf 0.171047.... 0.947469....
#> 4  0, 8.836.... 0.074318.... 0.993209....
#> 5  0, 10.15.... 0.078148.... 0.998369....
#> 6  0, 5.741.... 0.095427.... 0.967199....
#> 7  0, 1.188.... 0.095427.... 0.998925....
#> 8  0, 7.637.... 0.095427.... 0.994214....
#> 9        0, Inf 0.301436.... 0.727689....
#> 10       0, Inf 0.363336.... 0.950034....
x[[1]][["hairpins"]] # hairpin loops
#>   begin_low begin_high end_low end_high width_low width_high length_low
#> 1        -4          0      -3        4        72         81          8
#> 2         9          9      49       51        16         16          5
#> 3        14         14      23       25        37         38          5
#> 4        24         25      29       31        21         21          7
#>   length_high           dG   prevalence   background
#> 1          13 -Inf, -1.... 0.159193.... 0.014745....
#> 2           6 -Inf, -3.... 0.736407.... 0.006730....
#> 3           6 -Inf, -3.... 0.155095.... 0.013687....
#> 4           9 -Inf, -8.... 0.302477.... 0.001134....
head(x[[1]][["kmers"]]) # k-mer frequencies
#> [1]  189  872  435  151  113 1742

data(NonCodingRNA_Bacteria)
x <- NonCodingRNA_Bacteria
print(x)
#> $`tRNA-Ala`
#> NonCoding object with 10 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Arg`
#> NonCoding object with 7 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-Asn`
#> NonCoding object with 9 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Asp`
#> NonCoding object with 9 motifs, 2 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Cys`
#> NonCoding object with 8 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Gln`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Glu`
#> NonCoding object with 8 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Gly`
#> NonCoding object with 9 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-His`
#> NonCoding object with 10 motifs, 2 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Ile`
#> NonCoding object with 8 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Leu`
#> NonCoding object with 11 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-Lys`
#> NonCoding object with 9 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Met`
#> NonCoding object with 11 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Phe`
#> NonCoding object with 9 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Pro`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Ser`
#> NonCoding object with 10 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-Thr`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Trp`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Tyr`
#> NonCoding object with 14 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Val`
#> NonCoding object with 8 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Sec`
#> NonCoding object with 20 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`rRNA_5S-RF00001`
#> NonCoding object with 20 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`rRNA_16S-RF00177`
#> NonCoding object with 8 motifs, 20 hairpins, and 4-mer frequencies.
#> 
#> $`rRNA_23S-RF02541`
#> NonCoding object with 11 motifs, 20 hairpins, and 4-mer frequencies.
#> 
#> $`tmRNA-RF00023`
#> NonCoding object with 20 motifs, 6 hairpins, and 4-mer frequencies.
#> 
#> $`tmRNA_Alpha-RF01849`
#> NonCoding object with 17 motifs, 10 hairpins, and 4-mer frequencies.
#> 
#> $`RNase_P_class_A-RF00010`
#> NonCoding object with 20 motifs, 13 hairpins, and 4-mer frequencies.
#> 
#> $`RNase_P_class_B-RF00011`
#> NonCoding object with 15 motifs, 13 hairpins, and 4-mer frequencies.
#> 
#> $`SsrS-RF00013`
#> NonCoding object with 14 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`Intron_Gp_I-RF00028`
#> NonCoding object with 15 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`Intron_Gp_II-RF00029`
#> NonCoding object with 9 motifs, 2 hairpins, and 4-mer frequencies.
#> 
#> $`SmallSRP-RF00169`
#> NonCoding object with 16 motifs, 5 hairpins, and 4-mer frequencies.
#> 
#> $`Cyclic-di-GMP_Riboswitch-RF01051`
#> NonCoding object with 14 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`Cyclic-di-AMP_Riboswitch-RF00379`
#> NonCoding object with 19 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`T-box_Leader-RF00230`
#> NonCoding object with 20 motifs, 7 hairpins, and 4-mer frequencies.
#> 
#> $`Ribosomal_Protein_L10_Leader-RF00557`
#> NonCoding object with 11 motifs, 2 hairpins, and 4-mer frequencies.
#> 
#> $`Cobalamin_Riboswitch-RF00174`
#> NonCoding object with 11 motifs, 7 hairpins, and 4-mer frequencies.
#> 
#> $`TPP_Riboswitch-RF00059`
#> NonCoding object with 10 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`SAM_Riboswitch-RF00162`
#> NonCoding object with 14 motifs, 5 hairpins, and 4-mer frequencies.
#> 
#> $`Fluoride_Riboswitch-RF01734`
#> NonCoding object with 9 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`FMN_Riboswitch-RF00050`
#> NonCoding object with 12 motifs, 6 hairpins, and 4-mer frequencies.
#> 
#> $`Glycine_Riboswitch-RF00504`
#> NonCoding object with 10 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`HEARO-RF02033`
#> NonCoding object with 17 motifs, 3 hairpins, and 4-mer frequencies.
#> 
#> $`Flavo_1-RF01705`
#> NonCoding object with 8 motifs, 2 hairpins, and 4-mer frequencies.
#> 
#> $`Acido_Lenti_1-RF01687`
#> NonCoding object with 6 motifs, 2 hairpins, and 3-mer frequencies.
#> 
#> $`5'_ureB-RF02514`
#> NonCoding object with 10 motifs, 1 hairpin, and 4-mer frequencies.
#> 
class(x)
#> [1] "list"
attributes(x[[1]])
#> $names
#> [1] "motifs"       "hairpins"     "kmers"        "lengthScores" "dependence"  
#> 
#> $class
#> [1] "NonCoding"
#> 
#> $K
#> [1] 3
#> 
#> $minLength
#> [1] 47
#> 
#> $maxLength
#> [1] 154
#> 
#> $maxLoopLength
#> [1] 500
#> 
#> $background
#> meanlog   sdlog 
#>      NA      NA 
#> 
x[[1]] # the first non-coding RNA
#> NonCoding object with 10 motifs, 4 hairpins, and 3-mer frequencies.
x[[1]][["motifs"]] # sequence motifs
#>    begin_low begin_high end_low end_high          motif          pwm
#> 1          0          0      66       68       GGGGcydT 6.190340....
#> 2          8          8      58       60       AGCTCAGy 0.887309....
#> 3         16         16      55       57            TGG 0.005205....
#> 4         19         20      43       45   kAGAGCrCyTsc 0.024808....
#> 5         31         32      29       31 wTkGCAwksArGRG 0.261497....
#> 6         45         46      21       23       GTCrgsGG 0.016786....
#> 7         53         54      12       14      TTCGAwyCC 6.189574....
#> 8         62         63       1        3    scThrkCTCCA 6.190340....
#> 9         73         73       0        2              C 0.011661....
#> 10        74         74       0        0             CA 0.024090....
#>        minscore   prevalence   background
#> 1  0, 6.546.... 0.094171.... 0.990257....
#> 2  0, 7.992.... 0.122470.... 0.997144....
#> 3        0, Inf 0.171047.... 0.947469....
#> 4  0, 8.836.... 0.074318.... 0.993209....
#> 5  0, 10.15.... 0.078148.... 0.998369....
#> 6  0, 5.741.... 0.095427.... 0.967199....
#> 7  0, 1.188.... 0.095427.... 0.998925....
#> 8  0, 7.637.... 0.095427.... 0.994214....
#> 9        0, Inf 0.301436.... 0.727689....
#> 10       0, Inf 0.363336.... 0.950034....
x[[1]][["hairpins"]] # hairpin loops
#>   begin_low begin_high end_low end_high width_low width_high length_low
#> 1        -4          0      -3        4        72         81          8
#> 2         9          9      49       51        16         16          5
#> 3        14         14      23       25        37         38          5
#> 4        24         25      29       31        21         21          7
#>   length_high           dG   prevalence   background
#> 1          13 -Inf, -1.... 0.159193.... 0.014745....
#> 2           6 -Inf, -3.... 0.736407.... 0.006730....
#> 3           6 -Inf, -3.... 0.155095.... 0.013687....
#> 4           9 -Inf, -8.... 0.302477.... 0.001134....
head(x[[1]][["kmers"]]) # k-mer frequencies
#> [1]  189  872  435  151  113 1742