NonCoding Models for Common Non-Coding RNA Families

Pre-trained with NonCoding models for common RNA families found in genomes from organisms belonging to each domain of life.

Source

Models were built from sequences belonging to families in tRNADB-CE (http://trna.ie.niigata-u.ac.jp/cgi-bin/trnadb/index.cgi) or Rfam (http://rfam.xfam.org).

Details

A set of NonCoding models contained in a named list. Models were built from up to 1000 representative sequences per non-coding RNA family.

Examples

data(NonCodingRNA_Archaea)
data(NonCodingRNA_Bacteria)
data(NonCodingRNA_Eukarya)
names(NonCodingRNA_Bacteria)
#>  [1] "tRNA-Ala"                            
#>  [2] "tRNA-Arg"                            
#>  [3] "tRNA-Asn"                            
#>  [4] "tRNA-Asp"                            
#>  [5] "tRNA-Cys"                            
#>  [6] "tRNA-Gln"                            
#>  [7] "tRNA-Glu"                            
#>  [8] "tRNA-Gly"                            
#>  [9] "tRNA-His"                            
#> [10] "tRNA-Ile"                            
#> [11] "tRNA-Leu"                            
#> [12] "tRNA-Lys"                            
#> [13] "tRNA-Met"                            
#> [14] "tRNA-Phe"                            
#> [15] "tRNA-Pro"                            
#> [16] "tRNA-Ser"                            
#> [17] "tRNA-Thr"                            
#> [18] "tRNA-Trp"                            
#> [19] "tRNA-Tyr"                            
#> [20] "tRNA-Val"                            
#> [21] "tRNA-Sec"                            
#> [22] "rRNA_5S-RF00001"                     
#> [23] "rRNA_16S-RF00177"                    
#> [24] "rRNA_23S-RF02541"                    
#> [25] "tmRNA-RF00023"                       
#> [26] "tmRNA_Alpha-RF01849"                 
#> [27] "RNase_P_class_A-RF00010"             
#> [28] "RNase_P_class_B-RF00011"             
#> [29] "SsrS-RF00013"                        
#> [30] "Intron_Gp_I-RF00028"                 
#> [31] "Intron_Gp_II-RF00029"                
#> [32] "SmallSRP-RF00169"                    
#> [33] "Cyclic-di-GMP_Riboswitch-RF01051"    
#> [34] "Cyclic-di-AMP_Riboswitch-RF00379"    
#> [35] "T-box_Leader-RF00230"                
#> [36] "Ribosomal_Protein_L10_Leader-RF00557"
#> [37] "Cobalamin_Riboswitch-RF00174"        
#> [38] "TPP_Riboswitch-RF00059"              
#> [39] "SAM_Riboswitch-RF00162"              
#> [40] "Fluoride_Riboswitch-RF01734"         
#> [41] "FMN_Riboswitch-RF00050"              
#> [42] "Glycine_Riboswitch-RF00504"          
#> [43] "HEARO-RF02033"                       
#> [44] "Flavo_1-RF01705"                     
#> [45] "Acido_Lenti_1-RF01687"               
#> [46] "5'_ureB-RF02514"                     
head(NonCodingRNA_Bacteria)
#> $`tRNA-Ala`
#> NonCoding object with 10 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Arg`
#> NonCoding object with 7 motifs, 4 hairpins, and 4-mer frequencies.
#> 
#> $`tRNA-Asn`
#> NonCoding object with 9 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Asp`
#> NonCoding object with 9 motifs, 2 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Cys`
#> NonCoding object with 8 motifs, 4 hairpins, and 3-mer frequencies.
#> 
#> $`tRNA-Gln`
#> NonCoding object with 10 motifs, 3 hairpins, and 3-mer frequencies.
#>