Common Restriction Enzyme's Cut Sites

A character vector of common restriction sites named by the restriction enzyme that cuts at each site. Sequence specificity is listed in 5' to 3' orientation based on the IUPAC_CODE_MAP. The cut site is either signified by a ``/'' for palindromic sites, or two numbers giving the position of the top and bottom cut positions relative to the site's 3'-end.

Format

The format is: Named chr 1:224 "GACGT/C" "G/GTACC" "GT/MKAC" ...

attr(*, "names")= chr 1:224 "AatII" "Acc65I" "AccI" "AciI" ...

Source

Restriction enzymes sold by .

Examples

data(RESTRICTION_ENZYMES)
RESTRICTION_ENZYMES
#>              AatII             Acc65I               AccI               AciI 
#>          "GACGT/C"          "G/GTACC"          "GT/MKAC"      "CCGC(-3/-1)" 
#>               AclI               AcuI               AfeI              AflII 
#>          "AA/CGTT"    "CTGAAG(16/14)"          "AGC/GCT"          "C/TTAAG" 
#>             AflIII               AgeI               AhdI               AleI 
#>          "A/CRYGT"          "A/CCGGT"     "GACNNN/NNGTC"      "CACNN/NNGTG" 
#>               AluI               AlwI              AlwNI               ApaI 
#>            "AG/CT"       "GGATC(4/5)"       "CAGNNN/CTG"          "GGGCC/C" 
#>              ApaLI              ApeKI               ApoI               AscI 
#>          "G/TGCAC"           "G/CWGC"          "R/AATTY"        "GG/CGCGCC" 
#>               AseI              AsiSI               AvaI              AvaII 
#>          "AT/TAAT"        "GCGAT/CGC"          "C/YCGRG"           "G/GWCC" 
#>              AvrII              BaeGI              BamHI               BanI 
#>          "C/CTAGG"          "GKGCM/C"          "G/GATCC"          "G/GYRCC" 
#>              BanII               BbsI              BbvCI               BbvI 
#>          "GRGCY/C"      "GAAGAC(2/6)"   "CCTCAGC(-5/-2)"      "GCAGC(8/12)" 
#>               BccI              BceAI              BciVI               BclI 
#>       "CCATC(4/5)"     "ACGGC(12/14)"      "GTATCC(6/5)"          "T/GATCA" 
#>              BcoDI               BfaI              BfuAI              BfuCI 
#>       "GTCTC(1/5)"            "C/TAG"      "ACCTGC(4/8)"            "/GATC" 
#>               BglI              BglII               BlpI              BmgBI 
#>     "GCCNNNN/NGGC"          "A/GATCT"         "GC/TNAGC"    "CACGTC(-3/-3)" 
#>               BmrI               BmtI               BpmI             Bpu10I 
#>      "ACTGGG(5/4)"          "GCTAG/C"    "CTGGAG(16/14)"   "CCTNAGC(-5/-2)" 
#>              BpuEI              BsaAI              BsaBI              BsaHI 
#>    "CTTGAG(16/14)"          "YAC/GTR"      "GATNN/NNATC"          "GR/CGYC" 
#>               BsaI              BsaJI              BsaWI              BseRI 
#>      "GGTCTC(1/5)"          "C/CNNGG"          "W/CCGGW"     "GAGGAG(10/8)" 
#>              BseYI               BsgI              BsiEI            BsiHKAI 
#>    "CCCAGC(-5/-1)"    "GTGCAG(16/14)"          "CGRY/CG"          "GWGCW/C" 
#>              BsiWI               BslI              BsmBI              BsmFI 
#>          "C/GTACG"     "CCNNNNN/NNGG"      "CGTCTC(1/5)"     "GGGAC(10/14)" 
#>               BsmI           Bsp1286I             BspCNI              BspDI 
#>     "GAATGC(1/-1)"          "GDGCH/C"       "CTCAG(9/7)"          "AT/CGAT" 
#>              BspEI              BspHI              BspQI              BsrBI 
#>          "T/CCGGA"          "T/CATGA"     "GCTCTTC(1/4)"    "CCGCTC(-3/-3)" 
#>              BsrDI              BsrFI              BsrGI               BsrI 
#>      "GCAATG(2/0)"          "R/CCGGY"          "T/GTACA"      "ACTGG(1/-1)" 
#>             BssHII              BssKI              BssSI             BstAPI 
#>          "G/CGCGC"           "/CCNGG"    "CACGAG(-5/-1)"     "GCANNNN/NTGC" 
#>              BstBI             BstEII              BstNI              BstUI 
#>          "TT/CGAA"         "G/GTNACC"           "CC/WGG"            "CG/CG" 
#>              BstXI              BstYI            BstZ17I             Bsu36I 
#>    "CCANNNNN/NTGG"          "R/GATCY"          "GTA/TAC"         "CC/TNAGG" 
#>               BtgI              BtgZI              BtsCI               BtsI 
#>          "C/CRYGG"    "GCGATG(10/14)"       "GGATG(2/0)"      "GCAGTG(2/0)" 
#>           BtsIMutI              Cac8I             CviAII            CviKI-1 
#>       "CAGTG(2/0)"          "GCN/NGC"            "C/ATG"            "RG/CY" 
#>              CviQI               DdeI               DraI             DraIII 
#>            "G/TAC"           "C/TNAG"          "TTT/AAA"       "CACNNN/GTG" 
#>               DrdI               EaeI               EagI               EarI 
#>    "GACNNNN/NNGTC"          "Y/GGCCR"          "C/GGCCG"      "CTCTTC(1/4)" 
#>               EciI            Eco53kI              EcoNI           EcoO109I 
#>     "GGCGGA(11/9)"          "GAG/CTC"     "CCTNN/NNNAGG"         "RG/GNCCY" 
#>            EcoP15I              EcoRI              EcoRV               FatI 
#>    "CAGCAG(25/27)"          "G/AATTC"          "GAT/ATC"            "/CATG" 
#>               FauI             Fnu4HI               FokI               FseI 
#>       "CCCGC(4/6)"           "GC/NGC"      "GGATG(9/13)"        "GGCCGG/CC" 
#>               FspI              HaeII             HaeIII               HgaI 
#>          "TGC/GCA"          "RGCGC/Y"            "GG/CC"      "GACGC(5/10)" 
#>               HhaI             HincII            HindIII              HinfI 
#>            "GCG/C"          "GTY/RAC"          "A/AGCTT"           "G/ANTC" 
#>             HinP1I               HpaI               HphI           Hpy166II 
#>            "G/CGC"          "GTT/AAC"       "GGTGA(8/7)"          "GTN/NAC" 
#>            Hpy188I          Hpy188III             Hpy99I              HpyAV 
#>           "TCN/GA"          "TC/NNGA"           "CGWCG/"       "CCTTC(6/5)" 
#>          HpyCH4III           HpyCH4IV            HpyCH4V               KasI 
#>           "ACN/GT"            "A/CGT"            "TG/CA"          "G/GCGCC" 
#>               KpnI              MboII               MfeI              MluCI 
#>          "GGTAC/C"       "GAAGA(8/7)"          "C/AATTG"            "/AATT" 
#>               MluI               MlyI               MmeI               MnlI 
#>          "A/CGCGT"       "GAGTC(5/5)"    "TCCRAC(20/18)"        "CCTC(7/6)" 
#>               MscI               MseI               MslI             MspA1I 
#>          "TGG/CCA"            "T/TAA"      "CAYNN/NNRTG"          "CMG/CKG" 
#>               MwoI               NaeI               NarI               NciI 
#>     "GCNNNNN/NNGC"          "GCC/GGC"          "GG/CGCC"           "CC/SGG" 
#>               NcoI               NdeI             NgoMIV               NheI 
#>          "C/CATGG"          "CA/TATG"          "G/CCGGC"          "G/CTAGC" 
#>             NlaIII              NlaIV            NmeAIII               NotI 
#>            "CATG/"          "GGN/NCC"    "GCCGAG(21/19)"        "GC/GGCCGC" 
#>               NruI               NsiI               NspI          Nt.BstNBI 
#>          "TCG/CGA"          "ATGCA/T"          "RCATG/Y"      "GAGTC(4/-5)" 
#>               PacI             PaeR7I               PciI              PflFI 
#>        "TTAAT/TAA"          "C/TCGAG"          "A/CATGT"       "GACN/NNGTC" 
#>              PflMI              PluTI               PmeI               PmlI 
#>     "CCANNNN/NTGG"          "GGCGC/C"        "GTTT/AAAC"          "CAC/GTG" 
#>              PpuMI              PshAI               PsiI              PspGI 
#>         "RG/GWCCY"      "GACNN/NNGTC"          "TTA/TAA"           "/CCWGG" 
#>             PspOMI              PspXI               PstI               PvuI 
#>          "G/GGCCC"        "VC/TCGAGB"          "CTGCA/G"          "CGAT/CG" 
#>              PvuII               RsaI              RsrII               SacI 
#>          "CAG/CTG"            "GT/AC"         "CG/GWCCG"          "GAGCT/C" 
#>              SacII               SalI             Sau96I               SbfI 
#>          "CCGC/GG"          "G/TCGAC"           "G/GNCC"        "CCTGCA/GG" 
#>               ScaI              ScrFI              SexAI              SfaNI 
#>          "AGT/ACT"           "CC/NGG"         "A/CCWGGT"       "GCATC(5/9)" 
#>               SfcI               SfiI               SfoI              SgrAI 
#>          "C/TRYAG"   "GGCCNNNN/NGGCC"          "GGC/GCC"        "CR/CCGGYG" 
#>               SmaI               SmlI              SnaBI               SpeI 
#>          "CCC/GGG"          "C/TYRAG"          "TAC/GTA"          "A/CTAGT" 
#>               SphI               SspI               StuI               StyI 
#>          "GCATG/C"          "AAT/ATT"          "AGG/CCT"          "C/CWWGG" 
#>               SwaI              TaqaI               TfiI             Tsp45I 
#>        "ATTT/AAAT"            "T/CGA"           "G/AWTC"           "/GTSAC" 
#>              TspMI              TspRI               XbaI               XcmI 
#>          "C/CCGGG"       "NNCASTGNN/"          "T/CTAGA" "CCANNNNN/NNNNTGG" 
#>               XmnI               ZraI 
#>      "GAANN/NNTTC"          "GAC/GTC"