Remove Gap Characters in Sequences

Removes gaps ("-" or "." characters) in a set of sequences, either deleting all gaps or only those shared by all sequences in the set.

Usage

RemoveGaps(
  myXStringSet,
  removeGaps = "all",
  includeMask = FALSE,
  processors = 1
)

Arguments

myXStringSet

An AAStringSet, DNAStringSet, or RNAStringSet object containing sequences.

removeGaps

Determines how gaps ("-" or "." characters) are removed in the sequences. This should be (an unambiguous abbreviation of) one of "none", "all" or "common".

includeMask

Logical specifying whether to consider the mask character ("+") as a gap.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

Value

An XStringSet of the same type as myXStringSet.

Details

The removeGaps argument controls which gaps are removed in myXStringSet. Setting removeGaps to "all" will remove all gaps in the input sequences, whereas setting removeGaps to "common" will remove only gaps that exist in the same position in every sequence. Therefore, the latter method will leave gaps in place that are not shared by every sequence, requiring that the sequences in myXStringSet all be the same length (i.e., be aligned). Setting removeGaps to "none" will simply return myXStringSet unaltered.

See also

Author

Erik Wright eswright@pitt.edu

Examples

dna <- DNAStringSet(c("ACT-G-", "AC--G-"))
dna
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     6 ACT-G-
#> [2]     6 AC--G-
RemoveGaps(dna, "all")
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     4 ACTG
#> [2]     3 ACG
RemoveGaps(dna, "common")
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     4 ACTG
#> [2]     4 AC-G