Score a Multiple Sequence Alignment

Calculates a score for a multiple sequence alignment based on either sum-of-pairs or sum-of-adjacent-pairs scoring.

Usage

ScoreAlignment(
  myXStringSet,
  method = "pairs",
  perfectMatch = 1,
  misMatch = 0,
  gapOpening = -7.5,
  gapExtension = -0.6,
  substitutionMatrix = NULL,
  structures = NULL,
  structureMatrix = NULL,
  includeTerminalGaps = FALSE,
  weight = 1
)

Arguments

myXStringSet

An AAStringSet, DNAStringSet, or RNAStringSet object of aligned sequences.

method

Character string indicating the method of scoring. This should be (an abbreviation of) one of "pairs" for sum-of-pairs or "adjacent" for sum-of-adjacent-pairs. (See details section below.)

perfectMatch

Numeric giving the reward for aligning two matching nucleotides in the alignment. Only used for DNAStringSet or RNAStringSet inputs.

misMatch

Numeric giving the cost for aligning two mismatched nucleotides in the alignment. Only used for DNAStringSet or RNAStringSet inputs.

gapOpening

Numeric giving the cost for opening and closing a gap in the alignment.

gapExtension

Numeric giving the cost for extending an open gap in the alignment.

substitutionMatrix

Either a substitution matrix representing the substitution scores for an alignment or the name of the amino acid substitution matrix to use in alignment. The latter may be one of the following: BLOSUM45'', BLOSUM50'', BLOSUM62'', BLOSUM80'', BLOSUM100'', PAM30'', PAM40'', PAM70'', PAM120'', PAM250'', or ``MIQS''. The default (NULL) will use the perfectMatch and misMatch penalties for DNA/RNA or PFASUM50 for AA.

structures

Either NULL (the default) or a list of matrices with one list element per sequence in myXStringSet.

structureMatrix

A structure matrix if structures are supplied, or NULL otherwise.

includeTerminalGaps

Logical specifying whether or not to include terminal gaps ("-" or "." characters on each end of the sequence) into the calculation of score.

weight

A numeric vector of weights for each sequence, or a single number implying equal weights.

Value

A single numeric score.

Details

Sum-of-pairs scoring is the standard way to judge whether a set of sequences are homologous. ScoreAlignment calculates the sum-of-pairs score for myXStringSet when method is "pairs". This score can also be used to compare among different alignments of the same sequences. If method is "adjacent" then the sum-of-adjacent-pairs scores is calculated, where each sequence is compared to the next sequence. Hence, the input order of sequences in myXStringSet matters when method is "adjacent".

Both scores are linearly related to the number of sequences in the alignment and the number of sites in the alignment. Therefore, it is possible to normalize the score by dividing by the width and length (minus 1) of the myXStringSet.

See also

Author

Erik Wright eswright@pitt.edu

Examples

# small example
x <- DNAStringSet(c("C-G", "CTG", "C-G", "CTG"))
ScoreAlignment(x, method="pairs", gapOpening=-1) # +3 -1 +3 = 5
#> [1] 5
ScoreAlignment(x, method="adjacent", gapOpening=-1) # +3 -3 +3 = 3
#> [1] 3

# DNA alignment with the defaults
fas <- system.file("extdata", "Streptomyces_ITS_aligned.fas", package="DECIPHER")
dna <- readDNAStringSet(fas)
dna # input alignment
#> DNAStringSet object of length 88:
#>      width seq                                              names               
#>  [1]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTTTCCGAATGGGGAAACC supercont3.1 of S...
#>  [2]   627 NNNNCACACCGCCCGTCA-CGTC...GGGGTTTCCGAATGGGGAAACC supercont3.1 of S...
#>  [3]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTTTCCGAATGGGGAAACC supercont1.1 of S...
#>  [4]   627 CGTACACACCGCCCGTCA-CGTC...GGGGTTTCCGAATGGGGAAACC supercont1.1 of S...
#>  [5]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTTTCCGAATGGGGAAACC supercont1.1 of S...
#>  ...   ... ...
#> [84]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTTTCCGAATGGGGAAACC gi|297189896|ref|...
#> [85]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTGTCCGAATGGGGAAACC gi|224581106|ref|...
#> [86]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTGTCCGAATGGGGAAACC gi|224581106|ref|...
#> [87]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTGTCCGAATGGGGAAACC gi|224581106|ref|...
#> [88]   627 TGTACACACCGCCCGTCA-CGTC...GGGGTTTCCGAATGGGGAAACC gi|224581108|ref|...
ScoreAlignment(dna, method="pairs")
#> [1] 31679.96
ScoreAlignment(dna, method="adjacent")
#> [1] 41222.2

# provide a DNA substitution matrix for greater discerning power
sub <- matrix(c(1.5, -2.134, -0.739, -1.298,
    -2.134, 1.832, -2.462, 0.2,
    -0.739, -2.462, 1.522, -2.062,
    -1.298, 0.2, -2.062, 1.275),
  nrow=4,
  dimnames=list(DNA_BASES, DNA_BASES))
ScoreAlignment(dna, substitutionMatrix=sub)
#> [1] 46318.15

# use structures with an amino acid alignment
fas <- system.file("extdata", "50S_ribosomal_protein_L2.fas", package="DECIPHER")
dna <- readDNAStringSet(fas)
aa <- AlignTranslation(dna, type="AAStringSet")
#> Determining distance matrix based on shared 5-mers:
#> ================================================================================
#> 
#> Time difference of 0.73 secs
#> 
#> Clustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.02 secs
#> 
#> Aligning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.84 secs
#> 
#> Iteration 1 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.06 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.04 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.41 secs
#> 
#> Iteration 2 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.06 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.04 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.04 secs
#> 
structureMatrix <- matrix(c(0.187, -0.8, -0.873,
    -0.8, 0.561, -0.979,
    -0.873, -0.979, 0.221),
  nrow=3,
  dimnames=list(c("H", "E", "C"), c("H", "E", "C")))
ScoreAlignment(aa,
  structures=PredictHEC(aa, type="probabilities"),
  structureMatrix=structureMatrix)
#> [1] 270542.2