Produce a Staggered Alignment

Staggers overlapping characters in a multiple sequence alignment that are better explained by multiple insertions than multiple deletions.

Usage

StaggerAlignment(
  myXStringSet,
  tree = NULL,
  threshold = 3,
  fullLength = FALSE,
  processors = 1,
  verbose = TRUE
)

Arguments

myXStringSet

An AAStringSet, DNAStringSet, or RNAStringSet object of aligned sequences.

tree

A bifurcating dendrogram representing the evolutionary relationships between sequences, such as that created by TreeLine. The root should be the topmost node of the tree. The default (NULL) will automatically infer a tree from myXStringSet.

threshold

Numeric giving the ratio of insertions to deletions that must be met to stagger a region of the alignment. Specifically, the number of insertions divided by deletions must be less than threshold to stagger.

fullLength

Logical specifying whether the sequences are full-length (TRUE), or terminal gaps should be treated as missing data (FALSE, the default). Either a single logical, a vector with one logical per sequence, or a list with right and left components containing logicals for the right and left sides of the alignment.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to display progress.

Value

An XStringSet of aligned sequences.

Details

Multiple sequence aligners typically maximize true homologies at the expense of increased false homologies. StaggerAlignment creates a ``staggered alignment'' which separates regions of the alignment that are likely not homologous into separate regions. This re-balances the trade-off between true positives and false positives by decreasing the number of false homologies at the loss of some true homologies. The resulting alignment is less aesthetically pleasing because it is widened by the introduction of many gaps. However, in an evolutionary sense a staggered alignment is more correct because each aligned position represents a hypothesis about evolutionary events: overlapping characters between any two sequences represent positions common to their ancestor sequence that may have evolved through substitution.

The single parameter threshold controls the degree of staggering. Its value represents the ratio of insertions to deletions that must be crossed in order to stagger a region. A threshold of 1 would mean any region that could be better explained by separate insertions than deletions should be staggered. A higher value for threshold makes it more likely to stagger, and vise-versa. A very high value would conservatively stagger most regions with gaps, resulting in few false homologies but also fewer true homologies. The default value (3) is intended to remove more false homologies than it eliminates in true homologies. It may be preferable to tailor the threshold depending on the purpose of the alignment, as some downstream procedures (such as tree building) may be more or less sensitive to false homologies.

References

Coming soon!

See also

Author

Erik Wright eswright@pitt.edu

Examples

db <- system.file("extdata", "Bacteria_175seqs.sqlite", package="DECIPHER")
dna <- SearchDB(db, remove="all")
#> Search Expression:
#> select row_names, sequence from _Seqs where row_names in (select row_names
#> from Seqs)
#> 
#> DNAStringSet of length: 175
#> Time difference of 0.01 secs
#> 
alignedDNA <- AlignSeqs(dna)
#> Determining distance matrix based on shared 9-mers:
#> ================================================================================
#> 
#> Time difference of 1.52 secs
#> 
#> Clustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.03 secs
#> 
#> Aligning Sequences:
#> ================================================================================
#> 
#> Time difference of 1.41 secs
#> 
#> Iteration 1 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.07 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.02 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 1.06 secs
#> 
#> Iteration 2 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.07 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.38 secs
#> 
staggerDNA <- StaggerAlignment(alignedDNA)
#> Calculating distance matrix:
#> ================================================================================
#> 
#> Time difference of 0.06 secs
#> 
#> Constructing neighbor-joining tree:
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
#> Staggering insertions and deletions:
#> ================================================================================
#> 
#> Time difference of 1.67 secs
#> 
BrowseSeqs(staggerDNA, highlight=1)