Write Genes to a File

Writes predicted genes to a file in GenBank (gbk) or general feature format (gff).

Usage

WriteGenes(x, file = "", format = "gbk", append = FALSE)

Arguments

x

An object of class Genes.

file

A connection or a character string naming the file path where the tree should be exported. If "" (the default), the tree is printed to the standard output connection, the console unless redirected by sink.

format

Character specifying "gbk" or "gff" output format.

append

Logical indicating whether to append to an existing file. Only applicable if file is a character string. If FALSE (the default), then the file is overwritten.

Value

NULL.

Details

WriteGenes will write a "Genes" object to a GenBank (if format is "gbk") or general feature format (if format is "gff") file.

See also

Author

Erik Wright eswright@pitt.edu

Examples

# import a test genome
fas <- system.file("extdata",
  "Chlamydia_trachomatis_NC_000117.fas.gz",
  package="DECIPHER")
genome <- readDNAStringSet(fas)

x <- FindGenes(genome)
#> Iter  Models Start Motif  Init  Fold UpsNt  Term   RBS  Auto  Stop Genes
#> 1          1 18.40                 _                                 8861          1 18.40                 \                                 8861          1 18.40                 |                                 8861          1 18.40                 /                                 8861          1 18.40                 _                                 8861          1 18.40                 \                                 8861          1 18.40                 |                                 8861          1 18.40                 /                                 8861          1 18.40                 _                                 8862         12 18.40  0.48  0.90                                       8883         15 18.38  0.53  1.16  0.40  1.41                           8914         15 15.59  0.52  1.27  0.47  1.75  0.24                     8935         12 15.62  0.52  1.25  0.48  1.83  0.25  0.98  0.10  0.07   8966         16 15.61  0.51  1.24  0.51  1.91  0.25  1.08  0.10  0.07   8967         14 15.61  0.51  1.24  0.48  1.94  0.25  1.08  0.10  0.07   8968         13 15.60  0.52  1.25  0.53  1.97  0.26  1.07  0.10  0.07   8979         12 15.59  0.52  1.25  0.49  1.99  0.26  1.09  0.10  0.07   89610        12 15.59  0.52  1.25  0.49  1.99  0.26  1.13  0.10  0.07   89511        12 15.59  0.52  1.24  0.50  2.00  0.26  1.09  0.10  0.07   89512        12 15.59  0.52  1.24  0.51  2.00  0.26  1.12  0.10  0.07   894
#> 
#> Time difference of 29.73 secs
#> 
WriteGenes(x[1:10,], format="gbk")
#> DEFINITION  NC_000117.1 Chlamydia trachomatis D/UW-3/CX chromosome, complete genome
#> FEATURES             Location/Qualifiers
#>      CDS             1..1176
#>                      /note="ID=1_1;"
#>      CDS             complement(1321..1593)
#>                      /note="ID=1_2;"
#>      CDS             1794..2096
#>                      /note="ID=1_3;"
#>      CDS             2108..3583
#>                      /note="ID=1_4;"
#>      CDS             3585..5051
#>                      /note="ID=1_5;"
#>      CDS             complement(5150..6241)
#>                      /note="ID=1_6;"
#>      CDS             complement(6369..6938)
#>                      /note="ID=1_7;"
#>      CDS             7251..8201
#>                      /note="ID=1_8;"
#>      CDS             complement(8217..9119)
#>                      /note="ID=1_9;"
#>      CDS             9373..9804
#>                      /note="ID=1_10;"
WriteGenes(x[1:10,], format="gff")
#> ##gff-version 3
#> ##sequence-region NC_000117.1 1 1042519
#> NC_000117.1	DECIPHER_v2.29.1	CDS	1	1176	100	+	0	ID=1_1
#> NC_000117.1	DECIPHER_v2.29.1	CDS	1321	1593	100	-	0	ID=1_2
#> NC_000117.1	DECIPHER_v2.29.1	CDS	1794	2096	100	+	0	ID=1_3
#> NC_000117.1	DECIPHER_v2.29.1	CDS	2108	3583	100	+	0	ID=1_4
#> NC_000117.1	DECIPHER_v2.29.1	CDS	3585	5051	100	+	0	ID=1_5
#> NC_000117.1	DECIPHER_v2.29.1	CDS	5150	6241	100	-	0	ID=1_6
#> NC_000117.1	DECIPHER_v2.29.1	CDS	6369	6938	100	-	0	ID=1_7
#> NC_000117.1	DECIPHER_v2.29.1	CDS	7251	8201	100	+	0	ID=1_8
#> NC_000117.1	DECIPHER_v2.29.1	CDS	8217	9119	100	-	0	ID=1_9
#> NC_000117.1	DECIPHER_v2.29.1	CDS	9373	9804	100	+	0	ID=1_10