Change in Enthalpy of Hybridization of DNA/DNA Quadruplets in Solution

An 8D array with four adjacent base pairs of the DNA duplex. Each dimension has five elements defining the nucleotide at that position ("A", "C", "G", "T", or "-"). The array contains the standard enthalpy change of probe binding (dH, kcal/mol) for every quadruple base pairing.

Format

The format is: num 1:5, 1:5, 1:5, 1:5, 1:5, 1:5, 1:5, 1:5 -7.97 0 0 0 0 ... - attr(*, "dimnames")=List of 8 ..$ : chr 1:5 "A" "C" "G" "T" ... ..$ : chr 1:5 "A" "C" "G" "T" ... ..$ : chr 1:5 "A" "C" "G" "T" ... ..$ : chr 1:5 "A" "C" "G" "T" ... ..$ : chr 1:5 "A" "C" "G" "T" ... ..$ : chr 1:5 "A" "C" "G" "T" ... ..$ : chr 1:5 "A" "C" "G" "T" ... ..$ : chr 1:5 "A" "C" "G" "T" ...

Source

Data from a variety of publications by SantaLucia et al.

Details

The first four dimensions correspond to the four top strand positions from 5' to 3'. The fifth to eighth dimensions correspond to the four bottom strand positions from 5' to 3'.

References

SantaLucia, J., Jr., & Hicks, D. (2004) The Thermodynamics of DNA Structural Motifs. Annual Review of Biophysics and Biomolecular Structure, 33(1), 415-440. doi:10.1146/annurev.biophys.32.110601.141800.

Examples

data(deltaHrules)
# dH of the duplex AGCT / A-CT pairing
deltaHrules["A", "G", "C", "T", "A", "-", "C", "T"]
#> [1] -8.366159